requirements_pip.txt
gitpython
flask-sqlalchemy